refactor(credentials): break import cycle and defer keyring probe
- Move `ensureApifyDirectory` from utils.ts to files.ts to break the
  credentials.ts <-> utils.ts circular import.
- Drop the synthetic write/read/delete probe at backend selection time
  and treat the first real keyring write as the probe. On failure,
  downgrade to the file backend and persist the marker so future runs
  skip the keyring entirely. Avoids a duplicate macOS Keychain prompt
  on first login.
- Reword the "Corrupted local user info" error to accurately describe
  the stale-keyring-without-metadata state.
